@@ -8,40 +8,15 @@
 import Foundation
 import MapKit
 
-public class AppleMapViewFactory: NSObject, FlutterPlatformViewFactory {
-    
-    var registrar: FlutterPluginRegistrar
-    
-    public init(withRegistrar registrar: FlutterPluginRegistrar){
-        self.registrar = registrar
-        super.init()
-    }
-    
-    public func create(withFrame frame: CGRect, viewIdentifier viewId: Int64, arguments args: Any?) -> FlutterPlatformView {
-        let argsDictionary =  args as! Dictionary<String, Any>
-        
-        return AppleMapController(withFrame: frame, withRegistrar: registrar, withargs: argsDictionary, withId: viewId)
-        
-    }
-    
-    public func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
-        return FlutterStandardMessageCodec(readerWriter: FlutterStandardReaderWriter())
-    }
-}
-
-
-public class AppleMapController : NSObject, FlutterPlatformView, MKMapViewDelegate {
-    var mapView: FlutterMapView!
+public class AppleMapController: NSObject, FlutterPlatformView {
+    var mapView: FlutterMapView
     var registrar: FlutterPluginRegistrar
     var channel: FlutterMethodChannel
-    var annotationController: AnnotationController
-    var polylineController: PolylineController
-    var polygonController: PolygonController
-    var circleController: CircleController
     var initialCameraPosition: [String: Any]
     var options: [String: Any]
-    var onCalloutTapGestureRecognizer: UITapGestureRecognizer?
     var currentlySelectedAnnotation: String?
+    var snapShotOptions: MKMapSnapshotter.Options = MKMapSnapshotter.Options()
+    var snapShot: MKMapSnapshotter?
     
     public init(withFrame frame: CGRect, withRegistrar registrar: FlutterPluginRegistrar, withargs args: Dictionary<String, Any> ,withId id: Int64) {
         self.options = args["options"] as! [String: Any]
@@ -50,213 +25,215 @@ public class AppleMapController : NSObject, FlutterPlatformView, MKMapViewDelega
         self.mapView = FlutterMapView(channel: channel, options: options)
         self.registrar = registrar
         
-        self.annotationController = AnnotationController(mapView: mapView, channel: channel, registrar: registrar)
-        self.polylineController = PolylineController(mapView: mapView, channel: channel, registrar: registrar)
-        self.polygonController = PolygonController(mapView: mapView, channel: channel, registrar: registrar)
-        self.circleController = CircleController(mapView: mapView, channel: channel, registrar: registrar)
         self.initialCameraPosition = args["initialCameraPosition"]! as! Dictionary<String, Any>
         
         super.init()
         
         self.mapView.delegate = self
+        
         self.mapView.setCenterCoordinate(initialCameraPosition, animated: false)
         self.setMethodCallHandlers()
         
         if let annotationsToAdd: NSArray = args["annotationsToAdd"] as? NSArray {
-            self.annotationController.annotationsToAdd(annotations: annotationsToAdd)
+            self.annotationsToAdd(annotations: annotationsToAdd)
         }
         if let polylinesToAdd: NSArray = args["polylinesToAdd"] as? NSArray {
-            self.polylineController.addPolylines(polylineData: polylinesToAdd)
+            self.addPolylines(polylineData: polylinesToAdd)
         }
         if let polygonsToAdd: NSArray = args["polygonsToAdd"] as? NSArray {
-            self.polygonController.addPolygons(polygonData: polygonsToAdd)
+            self.addPolygons(polygonData: polygonsToAdd)
         }
         if let circlesToAdd: NSArray = args["circlesToAdd"] as? NSArray {
-            self.circleController.addCircles(circleData: circlesToAdd)
+            self.addCircles(circleData: circlesToAdd)
         }
-        
-        self.onCalloutTapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(self.calloutTapped(_:)))
+    }
+    
+    deinit {
+        self.removeAllAnnotations()
+        self.removeAllCircles()
+        self.removeAllPolygons()
+        self.removeAllPolylines()
     }
     
     public func view() -> UIView {
         return mapView
     }
     
-    // onIdle
-    public func mapView(_ mapView: MKMapView, regionDidChangeAnimated animated: Bool) {
-        self.channel.invokeMethod("camera#onIdle", arguments: "")
-    }
-    
-    // onMoveStarted
-    public func mapView(_ mapView: MKMapView, regionWillChangeAnimated animated: Bool) {
-        self.channel.invokeMethod("camera#onMoveStarted", arguments: "")
-    }
-    
-    public func mapView(_ mapView: MKMapView, didSelect view: MKAnnotationView)  {
-        if let annotation :FlutterAnnotation = view.annotation as? FlutterAnnotation  {
-            if annotation.infoWindowConsumesTapEvents {
-                view.addGestureRecognizer(self.onCalloutTapGestureRecognizer!)
-            }
-            self.currentlySelectedAnnotation = annotation.id
-            self.annotationController.onAnnotationClick(annotation: annotation)
-        }
-    }
-    
-    public func mapView(_ mapView: MKMapView, didDeselect view: MKAnnotationView) {
-        self.currentlySelectedAnnotation = nil
-        view.removeGestureRecognizer(self.onCalloutTapGestureRecognizer!)
-    }
-
-    
-    public func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-        if annotation is MKUserLocation {
-            return nil
-        } else if let flutterAnnotation = annotation as? FlutterAnnotation {
-            return self.annotationController.getAnnotationView(annotation: flutterAnnotation)
-        }
-        return nil
-    }
-    
-    public func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
-        if overlay is FlutterPolyline {
-            return polylineController.polylineRenderer(overlay: overlay)
-        } else if overlay is FlutterPolygon {
-            return polygonController.polygonRenderer(overlay: overlay)
-        } else if overlay is FlutterCircle {
-            return circleController.circleRenderer(overlay: overlay)
-        }
-        return MKOverlayRenderer()
-    }
-    
-    @objc func calloutTapped(_ sender: UITapGestureRecognizer? = nil) {
-        if self.currentlySelectedAnnotation != nil {
-            self.channel.invokeMethod("infoWindow#onTap", arguments: ["annotationId": self.currentlySelectedAnnotation!])
-        }
-    }
-    
     private func setMethodCallHandlers() {
-        channel.setMethodCallHandler({(call: FlutterMethodCall, result: FlutterResult) -> Void in
-            if let args :Dictionary<String, Any> = call.arguments as? Dictionary<String,Any> {
+        channel.setMethodCallHandler({ [unowned self] (call: FlutterMethodCall, result: @escaping FlutterResult) -> Void in
+            if let args: Dictionary<String, Any> = call.arguments as? Dictionary<String,Any> {
                 switch(call.method) {
                 case "annotations#update":
-                    if let annotationsToAdd = args["annotationsToAdd"] as? NSArray {
-                        if annotationsToAdd.count > 0 {
-                            self.annotationController.annotationsToAdd(annotations: annotationsToAdd)
-                        }
-                    }
-                    if let annotationsToChange = args["annotationsToChange"] as? NSArray {
-                        if annotationsToChange.count > 0 {
-                            self.annotationController.annotationsToChange(annotations: annotationsToChange)
-                        }
-                    }
-                    if let annotationsToDelete = args["annotationIdsToRemove"] as? NSArray {
-                        if annotationsToDelete.count > 0 {
-                            self.annotationController.annotationsIdsToRemove(annotationIds: annotationsToDelete)
-                        }
-                    }
+                    self.annotationUpdate(args: args)
                     result(nil)
+                    break
                 case "annotations#showInfoWindow":
-                    self.annotationController.showAnnotation(with: args["annotationId"] as! String)
+                    self.selectAnnotation(with: args["annotationId"] as! String)
+                    break
                 case "annotations#hideInfoWindow":
-                    self.annotationController.hideAnnotation(with: args["annotationId"] as! String)
+                    self.hideAnnotation(with: args["annotationId"] as! String)
+                    break
                 case "annotations#isInfoWindowShown":
-                    result(self.annotationController.isAnnotationSelected(with: args["annotationId"] as! String))
+                    result(self.isAnnotationSelected(with: args["annotationId"] as! String))
+                    break
                 case "polylines#update":
-                    if let polylinesToAdd: NSArray = args["polylinesToAdd"] as? NSArray {
-                        self.polylineController.addPolylines(polylineData: polylinesToAdd)
-                    }
-                    if let polylinesToChange: NSArray = args["polylinesToChange"] as? NSArray {
-                        self.polylineController.changePolylines(polylineData: polylinesToChange)
-                    }
-                    if let polylinesToRemove: NSArray = args["polylineIdsToRemove"] as? NSArray {
-                        self.polylineController.removePolylines(polylineIds: polylinesToRemove)
-                    }
-                    result(nil);
+                    self.polylineUpdate(args: args)
+                    result(nil)
+                    break
                 case "polygons#update":
-                    if let polyligonsToAdd: NSArray = args["polygonsToAdd"] as? NSArray {
-                        self.polygonController.addPolygons(polygonData: polyligonsToAdd)
-                    }
-                    if let polygonsToChange: NSArray = args["polygonsToChange"] as? NSArray {
-                        self.polygonController.changePolygons(polygonData: polygonsToChange)
-                    }
-                    if let polygonsToRemove: NSArray = args["polygonIdsToRemove"] as? NSArray {
-                        self.polygonController.removePolygons(polygonIds: polygonsToRemove)
-                    }
-                    result(nil);
+                    self.polygonUpdate(args: args)
+                    result(nil)
+                    break
                 case "circles#update":
-                    if let circlesToAdd: NSArray = args["circlesToAdd"] as? NSArray {
-                        self.circleController.addCircles(circleData: circlesToAdd)
-                    }
-                    if let circlesToChange: NSArray = args["circlesToChange"] as? NSArray {
-                        self.circleController.changeCircles(circleData: circlesToChange)
-                    }
-                    if let circlesToRemove: NSArray = args["circleIdsToRemove"] as? NSArray {
-                        self.circleController.removeCircles(circleIds: circlesToRemove)
-                    }
-                    result(nil);
+                    self.circleUpdate(args: args)
+                    result(nil)
+                    break
                 case "map#update":
                     self.mapView.interpretOptions(options: args["options"] as! Dictionary<String, Any>)
+                    break
                 case "camera#animate":
-                    let positionData :Dictionary<String, Any> = self.toPositionData(data: args["cameraUpdate"] as! Array<Any>, animated: true)
-                    if !positionData.isEmpty {
-                        guard let _ = positionData["moveToBounds"] else {
-                            self.mapView.setCenterCoordinate(positionData, animated: true)
-                            return
-                        }
-                        self.mapView.setBounds(positionData, animated: true)
-                    }
+                    self.animateCamera(args: args)
                     result(nil)
+                    break
                 case "camera#move":
-                    let positionData :Dictionary<String, Any> = self.toPositionData(data: args["cameraUpdate"] as! Array<Any>, animated: false)
-                    if !positionData.isEmpty {
-                        guard let _ = positionData["moveToBounds"] else {
-                            self.mapView.setCenterCoordinate(positionData, animated: false)
-                            return
-                        }
-                        self.mapView.setBounds(positionData, animated: false)
-                    }
+                    self.moveCamera(args: args)
                     result(nil)
+                    break
                 case "camera#convert":
-                    guard let annotation = args["annotation"] as? Array<Double> else {
-                        result(nil)
-                        return
-                    }
-                    let point = self.mapView.convert(CLLocationCoordinate2D(latitude: annotation[0] , longitude: annotation[1]), toPointTo: self.view())
-                    result(["point": [point.x, point.y]])
+                    self.cameraConvert(args: args, result: result)
+                    break
+                case "map#takeSnapshot":
+                    self.takeSnapshot(options: SnapshotOptions.init(options: args), onCompletion: { (snapshot: FlutterStandardTypedData?, error: Error?) -> Void in
+                        result(snapshot ?? error)
+                    })
                 default:
                     result(FlutterMethodNotImplemented)
-                    return
+                    break
                 }
             } else {
                 switch call.method {
                 case "map#getVisibleRegion":
                     result(self.mapView.getVisibleRegion())
+                    break
                 case "map#isCompassEnabled":
                     if #available(iOS 9.0, *) {
                         result(self.mapView.showsCompass)
                     } else {
                         result(false)
                     }
+                    break
                 case "map#isPitchGesturesEnabled":
                     result(self.mapView.isPitchEnabled)
+                    break
                 case "map#isScrollGesturesEnabled":
                     result(self.mapView.isScrollEnabled)
+                    break
                 case "map#isZoomGesturesEnabled":
                     result(self.mapView.isZoomEnabled)
+                    break
                 case "map#isRotateGesturesEnabled":
                     result(self.mapView.isRotateEnabled)
+                    break
                 case "map#isMyLocationButtonEnabled":
                     result(self.mapView.isMyLocationButtonShowing ?? false)
+                    break
                 case "map#getMinMaxZoomLevels":
                     result([self.mapView.minZoomLevel, self.mapView.maxZoomLevel])
+                    break
                 case "camera#getZoomLevel":
                     result(self.mapView.calculatedZoomLevel)
+                    break
                 default:
                     result(FlutterMethodNotImplemented)
-                    return
+                    break
                 }
             }
         })
+    }
+    
+    private func annotationUpdate(args: Dictionary<String, Any>) -> Void {
+        if let annotationsToAdd = args["annotationsToAdd"] as? NSArray {
+            if annotationsToAdd.count > 0 {
+                self.annotationsToAdd(annotations: annotationsToAdd)
+            }
+        }
+        if let annotationsToChange = args["annotationsToChange"] as? NSArray {
+            if annotationsToChange.count > 0 {
+                self.annotationsToChange(annotations: annotationsToChange)
+            }
+        }
+        if let annotationsToDelete = args["annotationIdsToRemove"] as? NSArray {
+            if annotationsToDelete.count > 0 {
+                self.annotationsIdsToRemove(annotationIds: annotationsToDelete)
+            }
+        }
+    }
+    
+    private func polygonUpdate(args: Dictionary<String, Any>) -> Void {
+        if let polyligonsToAdd: NSArray = args["polygonsToAdd"] as? NSArray {
+            self.addPolygons(polygonData: polyligonsToAdd)
+        }
+        if let polygonsToChange: NSArray = args["polygonsToChange"] as? NSArray {
+            self.changePolygons(polygonData: polygonsToChange)
+        }
+        if let polygonsToRemove: NSArray = args["polygonIdsToRemove"] as? NSArray {
+            self.removePolygons(polygonIds: polygonsToRemove)
+        }
+    }
+    
+    private func polylineUpdate(args: Dictionary<String, Any>) -> Void {
+        if let polylinesToAdd: NSArray = args["polylinesToAdd"] as? NSArray {
+            self.addPolylines(polylineData: polylinesToAdd)
+        }
+        if let polylinesToChange: NSArray = args["polylinesToChange"] as? NSArray {
+            self.changePolylines(polylineData: polylinesToChange)
+        }
+        if let polylinesToRemove: NSArray = args["polylineIdsToRemove"] as? NSArray {
+            self.removePolylines(polylineIds: polylinesToRemove)
+        }
+    }
+    
+    private func circleUpdate(args: Dictionary<String, Any>) -> Void {
+        if let circlesToAdd: NSArray = args["circlesToAdd"] as? NSArray {
+            self.addCircles(circleData: circlesToAdd)
+        }
+        if let circlesToChange: NSArray = args["circlesToChange"] as? NSArray {
+            self.changeCircles(circleData: circlesToChange)
+        }
+        if let circlesToRemove: NSArray = args["circleIdsToRemove"] as? NSArray {
+            self.removeCircles(circleIds: circlesToRemove)
+        }
+    }
+    
+    private func moveCamera(args: Dictionary<String, Any>) -> Void {
+        let positionData: Dictionary<String, Any> = self.toPositionData(data: args["cameraUpdate"] as! Array<Any>, animated: true)
+        if !positionData.isEmpty {
+            guard let _ = positionData["moveToBounds"] else {
+                self.mapView.setCenterCoordinate(positionData, animated: false)
+                return
+            }
+            self.mapView.setBounds(positionData, animated: false)
+        }
+    }
+    
+    private func animateCamera(args: Dictionary<String, Any>) -> Void {
+        let positionData: Dictionary<String, Any> = self.toPositionData(data: args["cameraUpdate"] as! Array<Any>, animated: true)
+        if !positionData.isEmpty {
+            guard let _ = positionData["moveToBounds"] else {
+                self.mapView.setCenterCoordinate(positionData, animated: true)
+                return
+            }
+            self.mapView.setBounds(positionData, animated: true)
+        }
+    }
+    
+    private func cameraConvert(args: Dictionary<String, Any>, result: FlutterResult) -> Void {
+        guard let annotation = args["annotation"] as? Array<Double> else {
+            result(nil)
+            return
+        }
+        let point = self.mapView.convert(CLLocationCoordinate2D(latitude: annotation[0] , longitude: annotation[1]), toPointTo: self.view())
+        result(["point": [point.x, point.y]])
     }
     
     private func toPositionData(data: Array<Any>, animated: Bool) -> Dictionary<String, Any> {
@@ -299,5 +276,109 @@ public class AppleMapController : NSObject, FlutterPlatformView, MKMapViewDelega
             return positionData
         }
         return [:]
+    }
+}
+
+
+extension AppleMapController: MKMapViewDelegate {
+    // onIdle
+    public func mapView(_ mapView: MKMapView, regionDidChangeAnimated animated: Bool) {
+        self.channel.invokeMethod("camera#onIdle", arguments: "")
+    }
+    
+    // onMoveStarted
+    public func mapView(_ mapView: MKMapView, regionWillChangeAnimated animated: Bool) {
+        self.channel.invokeMethod("camera#onMoveStarted", arguments: "")
+    }
+    
+    public func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
+        if overlay is FlutterPolyline {
+            return self.polylineRenderer(overlay: overlay)
+        } else if overlay is FlutterPolygon {
+            return self.polygonRenderer(overlay: overlay)
+        } else if overlay is FlutterCircle {
+            return self.circleRenderer(overlay: overlay)
+        }
+        return MKOverlayRenderer()
+    }
+}
+
+extension AppleMapController {
+    private func takeSnapshot(options: SnapshotOptions, onCompletion: @escaping (FlutterStandardTypedData?, Error?) -> Void) {
+        // MKMapSnapShotOptions setting.
+        snapShotOptions.region = self.mapView.region
+        snapShotOptions.size = self.mapView.frame.size
+        snapShotOptions.scale = UIScreen.main.scale
+        snapShotOptions.showsBuildings = options.showBuildings
+        snapShotOptions.showsPointsOfInterest = options.showPointsOfInterest
+        
+        // Set MKMapSnapShotOptions to MKMapSnapShotter.
+        snapShot = MKMapSnapshotter(options: snapShotOptions)
+        
+        snapShot?.cancel()
+        
+        if #available(iOS 10.0, *) {
+            snapShot?.start { [unowned self] snapshot, error in
+                guard let snapshot = snapshot, error == nil else {
+                    onCompletion(nil, error)
+                    return
+                }
+
+                let image = UIGraphicsImageRenderer(size: self.snapShotOptions.size).image { context in
+                    snapshot.image.draw(at: .zero)
+                    let rect = snapShotOptions.mapRect
+                    if options.showAnnotations {
+                        for annotation in self.mapView.getMapViewAnnotations() {
+                            self.drawAnnotations(annotation: annotation, point: snapshot.point(for: annotation!.coordinate))
+                        }
+                    }
+                    if options.showOverlays {
+                        for overlay in self.mapView.overlays {
+                            if ((overlay.intersects?(rect)) != nil) {
+                                self.drawOverlays(overlay: overlay, snapshot: snapshot, context: context)
+                            }
+                        }
+                    }
+                }
+
+                if let imageData = image.pngData() {
+                    onCompletion(FlutterStandardTypedData.init(bytes: imageData), nil)
+                }
+            }
+        }
+    }
+    
+    private func drawAnnotations(annotation: FlutterAnnotation?, point: CGPoint) {
+        guard annotation != nil else {
+            return
+        }
+        let annotationView = self.getAnnotationView(annotation: annotation!)
+        
+        var offsetPoint = point
+        
+        offsetPoint.x -= annotationView.bounds.width / 2
+        offsetPoint.y -= annotationView.bounds.height / 2
+        
+        
+        if #available(iOS 11.0, *), annotationView is MKMarkerAnnotationView {
+            annotationView.drawHierarchy(in: CGRect(x: offsetPoint.x, y: offsetPoint.y, width: annotationView.bounds.width, height: annotationView.bounds.height), afterScreenUpdates: true)
+        } else {
+            offsetPoint.x += annotationView.centerOffset.x
+            offsetPoint.y += annotationView.centerOffset.y
+            let annotationImage = annotationView.image
+            annotationImage?.draw(at: offsetPoint)
+        }
+    }
+    
+    @available(iOS 10.0, *)
+    private func drawOverlays(overlay: MKOverlay?, snapshot: MKMapSnapshotter.Snapshot, context: UIGraphicsRendererContext) {
+        guard overlay != nil else {
+            return
+        }
+        
+        if let flutterOverlay: FlutterOverlay = overlay as? FlutterOverlay {
+            flutterOverlay.getCAShapeLayer(snapshot: snapshot).render(in: context.cgContext)
+        }
+        
     }
 }
