@@ -8,30 +8,8 @@
 import Foundation
 import MapKit
 
-class PolylineController {
+extension AppleMapController: PolylineDelegate {
     
-    let availableCaps: Dictionary<String, CGLineCap> = [
-        "buttCap": CGLineCap.butt,
-        "roundCap": CGLineCap.round,
-        "squareCap": CGLineCap.square
-    ]
-    
-    let availableJointTypes: Array<CGLineJoin> = [
-        CGLineJoin.miter,
-        CGLineJoin.bevel,
-        CGLineJoin.round
-    ]
-    
-    var mapView: MKMapView
-    var channel: FlutterMethodChannel
-    var registrar: FlutterPluginRegistrar
-
-    init(mapView: MKMapView, channel: FlutterMethodChannel, registrar: FlutterPluginRegistrar) {
-        self.mapView = mapView
-        self.channel = channel
-        self.registrar = registrar
-    }
-
     func polylineRenderer(overlay: MKOverlay) -> MKOverlayRenderer {
         // Make sure we are rendering a polyline.
         guard let polyline = overlay as? MKPolyline else {
@@ -43,13 +21,9 @@ class PolylineController {
             if flutterPolyline.isVisible! {
                 polylineRenderer.strokeColor = flutterPolyline.color
                 polylineRenderer.lineWidth = flutterPolyline.width ?? 1.0
-                polylineRenderer.lineDashPattern = linePatternToArray(patternData: flutterPolyline.pattern, lineWidth: flutterPolyline.width)
-                polylineRenderer.lineJoin = availableJointTypes[flutterPolyline.lineJoin ?? 2]
-                if flutterPolyline.pattern != nil && flutterPolyline.pattern?.count != 0 {
-                    polylineRenderer.lineCap = getLineCapForLinePattern(linePatternData: flutterPolyline.pattern)
-                } else {
-                    polylineRenderer.lineCap = availableCaps[flutterPolyline.capType ?? "buttCap"]!
-                }
+                polylineRenderer.lineDashPattern = flutterPolyline.pattern
+                polylineRenderer.lineJoin = flutterPolyline.lineJoin!
+                polylineRenderer.lineCap = flutterPolyline.capType!
             } else {
                 polylineRenderer.strokeColor = UIColor.clear
                 polylineRenderer.lineWidth = 0.0
@@ -67,7 +41,7 @@ class PolylineController {
     }
 
     func changePolylines(polylineData data: NSArray) {
-        let oldOverlays: [MKOverlay] = mapView.overlays
+        let oldOverlays: [MKOverlay] = self.mapView.overlays
         for oldOverlay in oldOverlays {
             if oldOverlay is FlutterPolyline {
                 let oldFlutterPolyline = oldOverlay as! FlutterPolyline
@@ -85,25 +59,33 @@ class PolylineController {
     }
 
     func removePolylines(polylineIds: NSArray) {
-        for overlay in mapView.overlays {
+        for overlay in self.mapView.overlays {
             if let polyline = overlay as? FlutterPolyline {
                 if polylineIds.contains(polyline.id!) {
-                    mapView.removeOverlay(polyline)
+                    self.mapView.removeOverlay(polyline)
                 }
             }
         }
     }
     
+    func removeAllPolylines() {
+        for overlay in self.mapView.overlays {
+            if let polyline = overlay as? FlutterPolyline {
+                self.mapView.removeOverlay(polyline)
+            }
+        }
+    }
+    
     private func updatePolylinesOnMap(oldPolyline: FlutterPolyline, newPolyline: FlutterPolyline) {
-        mapView.removeOverlay(oldPolyline)
+        self.mapView.removeOverlay(oldPolyline)
         addPolyline(polyline: newPolyline)
     }
     
     private func addPolyline(polyline: FlutterPolyline) {
         if polyline.zIndex == nil || polyline.zIndex == -1 {
-            mapView.addOverlay(polyline)
+            self.mapView.addOverlay(polyline)
         } else {
-            mapView.insertOverlay(polyline, at: polyline.zIndex ?? 0)
+            self.mapView.insertOverlay(polyline, at: polyline.zIndex ?? 0)
         }
     }
     
