@@ -143,7 +143,7 @@ class Annotation {
   /// * has no info window text; [infoWindowText] is `InfoWindowText.noText`
   /// * is positioned at 0, 0; [position] is `LatLng(0.0, 0.0)`
   /// * is visible; [visible] is true
-  const Annotation({
+  Annotation({
     required this.annotationId,
     this.alpha = 1.0,
     this.anchor = const Offset(0.5, 1.0),
@@ -153,9 +153,9 @@ class Annotation {
     this.position = const LatLng(0.0, 0.0),
     this.onTap,
     this.visible = true,
+    this.zIndex = -1,
     this.onDragEnd,
     this.rotation = 0.0,
-    this.zPosition = 0.0,
   }) : assert(0.0 <= alpha && alpha <= 1.0);
 
   /// Uniquely identifies a [Annotation].
@@ -198,8 +198,12 @@ class Annotation {
   /// Rotation of the marker image in degrees clockwise from the [anchor] point.
   final double rotation;
 
-  // zPosition of the Annotation
-  final double zPosition;
+  /// The z-index of the annotation, used to determine relative drawing order of
+  /// map overlays.
+  ///
+  /// Overlays are drawn in order of z-index, so that lower values means drawn
+  /// earlier, and thus appearing to be closer to the surface of the Earth.
+  double zIndex;
 
   /// Creates a new [Annotation] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
@@ -212,10 +216,10 @@ class Annotation {
     InfoWindow? infoWindowParam,
     LatLng? positionParam,
     bool? visibleParam,
+    double? zIndexParam,
     VoidCallback? onTapParam,
     ValueChanged<LatLng>? onDragEndParam,
     double? rotationParam,
-    double? zPositionParam,
   }) {
     return Annotation(
       annotationId: annotationId,
@@ -227,9 +231,9 @@ class Annotation {
       position: positionParam ?? position,
       onTap: onTapParam ?? onTap,
       visible: visibleParam ?? visible,
+      zIndex: zIndexParam ?? zIndex,
       onDragEnd: onDragEndParam ?? onDragEnd,
       rotation: rotationParam ?? rotation,
-      zPosition: zPositionParam ?? zPosition,
     );
   }
 
@@ -251,8 +255,8 @@ class Annotation {
     addIfPresent('visible', visible);
     addIfPresent('position', position._toJson());
     addIfPresent('rotation', rotation);
-    addIfPresent('zPosition', zPosition);
-    addIfPresent('anchor', anchor != null ? [anchor.dx, anchor.dy] : null);
+    addIfPresent('anchor', [anchor.dx, anchor.dy]);
+    addIfPresent('zIndex', zIndex);
     return json;
   }
 
@@ -261,7 +265,15 @@ class Annotation {
     if (identical(this, other)) return true;
     if (other is! Annotation) return false;
     final Annotation typedOther = other;
-    return annotationId == typedOther.annotationId;
+    return annotationId == typedOther.annotationId &&
+        alpha == typedOther.alpha &&
+        anchor == typedOther.anchor &&
+        draggable == typedOther.draggable &&
+        icon == typedOther.icon &&
+        infoWindow == typedOther.infoWindow &&
+        position == typedOther.position &&
+        visible == typedOther.visible &&
+        zIndex == typedOther.zIndex;
   }
 
   @override
@@ -269,9 +281,9 @@ class Annotation {
 
   @override
   String toString() {
-    return 'Annotation{annotationId: $annotationId, alpha: $alpha, draggable: $draggable,'
-        'icon: $icon, infoWindow: $infoWindow, position: $position ,visible: $visible,' 
-        'onTap: $onTap, rotation: $rotation, zPosition: $zPosition,}';
+    return 'Annotation{annotationId: $annotationId, alpha: $alpha, draggable: $draggable, '
+        'icon: $icon, infoWindow: $infoWindow, position: $position ,visible: $visible, '
+        'onTap: $onTap}, zIndex: $zIndex, rotation: $rotation, onTap: $onTap}';
   }
 }
 
