@@ -66,6 +66,10 @@ class AppleMapController {
       case 'infoWindow#onTap':
         _appleMapState.onInfoWindowTap(call.arguments['annotationId']);
         break;
+      case 'annotation#onZIndexChanged':
+        _appleMapState.onAnnotationZIndexChanged(
+            call.arguments['annotationId'], call.arguments['zIndex']);
+        break;
       case 'map#onTap':
         _appleMapState.onTap(LatLng._fromJson(call.arguments['position'])!);
         break;
@@ -194,7 +198,7 @@ class AppleMapController {
         <String, String>{'annotationId': annotationId.value});
   }
 
-  /// Changes the map camera position.
+  /// Changes the map camera position without animating the transition.
   ///
   /// The returned [Future] completes after the change has been made on the
   /// platform side.
@@ -232,5 +236,12 @@ class AppleMapController {
     }
     final doubles = List<double>.from(point?['point']);
     return Offset(doubles.first, doubles.last);
+  }
+
+  /// Returns the image bytes of the map
+  Future<Uint8List?> takeSnapshot(
+      [SnapshotOptions snapshotOptions = const SnapshotOptions()]) {
+    return channel.invokeMethod<Uint8List>(
+        'map#takeSnapshot', snapshotOptions._toMap());
   }
 }
